pyproject.toml
[tool.poetry]
authors = ["Matt Love <matt@hyperglass.dev>"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Topic :: Multimedia :: Graphics :: Graphics Conversion",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Software Development :: User Interfaces",
  "Typing :: Typed",
]
description = "Favicon generator for Python 3 with strongly typed sync & async APIs, CLI, & HTML generation."
homepage = "https://github.com/thatmattlove/favicons"
license = "BSD-3-Clause-Clear"
name = "favicons"
readme = "README.md"
repository = "https://github.com/thatmattlove/favicons"
version = "0.0.9"

[tool.poetry.dependencies]
pillow = ">=7.2,<9.0"
python = ">=3.6.1,<4.0"
rich = ">=6.0,<11.0"
svglib = "^1.0.0"
typer = "^0.3.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.3"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.0"
flake8-breakpoint = "^1.1.0"
flake8-bugbear = "^20.1.4"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.2.3"
flake8-deprecated = "^1.3"
flake8-docstrings = "^1.5.0"
flake8-eradicate = "^0.4.0"
flake8-if-expr = "^1.0.4"
flake8-isort = "^3.0.1"
flake8-pie = "^0.5.0"
flake8-plugin-utils = "^1.3.0"
flake8-polyfill = "^1.0.2"
flake8-print = "^3.1.4"
flake8-return = "^1.1.2"
isort = "^4.3"
mccabe = "^0.6.1"
mypy = "^0.782"
pep8-naming = "^0.11.1"
pre-commit = "^2.6.0"

[tool.poetry.scripts]
favicons = "favicons:cli"

[tool.black]
line-length = 100

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
